TST: Clean/moved unused/sparsely use fixtures
