+++ pandas/conftest.py
@@ -30,7 +30,6 @@ from datetime import (
 from decimal import Decimal
 import operator
 import os
-from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Callable,
@@ -775,23 +774,6 @@ def series_with_simple_index(index) -> Series:
     return _create_series(index)
 
 
-@pytest.fixture
-def series_with_multilevel_index() -> Series:
-    """
-    Fixture with a Series with a 2-level MultiIndex.
-    """
-    arrays = [
-        ["bar", "bar", "baz", "baz", "qux", "qux", "foo", "foo"],
-        ["one", "two", "one", "two", "one", "two", "one", "two"],
-    ]
-    tuples = zip(*arrays)
-    index = MultiIndex.from_tuples(tuples)
-    data = np.random.default_rng(2).standard_normal(8)
-    ser = Series(data, index=index)
-    ser.iloc[3] = np.nan
-    return ser
-
-
 _narrow_series = {
     f"{dtype.__name__}-series": tm.make_rand_series(name="a", dtype=dtype)
     for dtype in tm.NARROW_NP_DTYPES
@@ -866,35 +848,6 @@ def int_frame() -> DataFrame:
 
 
 @pytest.fixture
-def datetime_frame() -> DataFrame:
-    """
-    Fixture for DataFrame of floats with DatetimeIndex
-
-    Columns are ['A', 'B', 'C', 'D']
-
-                       A         B         C         D
-    2000-01-03 -1.122153  0.468535  0.122226  1.693711
-    2000-01-04  0.189378  0.486100  0.007864 -1.216052
-    2000-01-05  0.041401 -0.835752 -0.035279 -0.414357
-    2000-01-06  0.430050  0.894352  0.090719  0.036939
-    2000-01-07 -0.620982 -0.668211 -0.706153  1.466335
-    2000-01-10 -0.752633  0.328434 -0.815325  0.699674
-    2000-01-11 -2.236969  0.615737 -0.829076 -1.196106
-    ...              ...       ...       ...       ...
-    2000-02-03  1.642618 -0.579288  0.046005  1.385249
-    2000-02-04 -0.544873 -1.160962 -0.284071 -1.418351
-    2000-02-07 -2.656149 -0.601387  1.410148  0.444150
-    2000-02-08 -1.201881 -1.289040  0.772992 -1.445300
-    2000-02-09  1.377373  0.398619  1.008453 -0.928207
-    2000-02-10  0.473194 -0.636677  0.984058  0.511519
-    2000-02-11 -0.965556  0.408313 -1.312844 -0.381948
-
-    [30 rows x 4 columns]
-    """
-    return DataFrame(tm.getTimeSeriesData())
-
-
-@pytest.fixture
 def float_frame() -> DataFrame:
     """
     Fixture for DataFrame of floats with index of unique strings
@@ -921,24 +874,6 @@ def float_frame() -> DataFrame:
     [30 rows x 4 columns]
     """
     return DataFrame(tm.getSeriesData())
-
-
-@pytest.fixture
-def mixed_type_frame() -> DataFrame:
-    """
-    Fixture for DataFrame of float/int/string columns with RangeIndex
-    Columns are ['a', 'b', 'c', 'float32', 'int32'].
-    """
-    return DataFrame(
-        {
-            "a": 1.0,
-            "b": 2,
-            "c": "foo",
-            "float32": np.array([1.0] * 10, dtype="float32"),
-            "int32": np.array([1] * 10, dtype="int32"),
-        },
-        index=np.arange(10),
-    )
 
 
 @pytest.fixture
@@ -1175,16 +1110,6 @@ def strict_data_files(pytestconfig):
 
 
 @pytest.fixture
-def tests_path() -> Path:
-    return Path(__file__).parent / "tests"
-
-
-@pytest.fixture
-def tests_io_data_path(tests_path) -> Path:
-    return tests_path / "io" / "data"
-
-
-@pytest.fixture
 def datapath(strict_data_files: str) -> Callable[..., str]:
     """
     Get the path to a data file.
@@ -1216,14 +1141,6 @@ def datapath(strict_data_files: str) -> Callable[..., str]:
         return path
 
     return deco
-
-
-@pytest.fixture
-def iris(datapath) -> DataFrame:
-    """
-    The iris dataset as a DataFrame.
-    """
-    return pd.read_csv(datapath("io", "data", "csv", "iris.csv"))
 
 
 # ----------------------------------------------------------------
@@ -1903,28 +1820,6 @@ def sort_by_key(request):
     Tests None (no key) and the identity key.
     """
     return request.param
-
-
-@pytest.fixture()
-def fsspectest():
-    pytest.importorskip("fsspec")
-    from fsspec import register_implementation
-    from fsspec.implementations.memory import MemoryFileSystem
-    from fsspec.registry import _registry as registry
-
-    class TestMemoryFS(MemoryFileSystem):
-        protocol = "testmem"
-        test = [None]
-
-        def __init__(self, **kwargs) -> None:
-            self.test[0] = kwargs.pop("test", None)
-            super().__init__(**kwargs)
-
-    register_implementation("testmem", TestMemoryFS, clobber=True)
-    yield TestMemoryFS()
-    registry.pop("testmem", None)
-    TestMemoryFS.test[0] = None
-    TestMemoryFS.store.clear()
 
 
 @pytest.fixture(
