+++ pandas/tests/frame/methods/test_astype.py
@@ -67,9 +67,19 @@ class TestAstype:
         casted = mixed_float_frame.reindex(columns=["A", "B"]).astype("float16")
         _check_cast(casted, "float16")
 
-    def test_astype_mixed_type(self, mixed_type_frame):
+    def test_astype_mixed_type(self):
         # mixed casting
-        mn = mixed_type_frame._get_numeric_data().copy()
+        df = DataFrame(
+            {
+                "a": 1.0,
+                "b": 2,
+                "c": "foo",
+                "float32": np.array([1.0] * 10, dtype="float32"),
+                "int32": np.array([1] * 10, dtype="int32"),
+            },
+            index=np.arange(10),
+        )
+        mn = df._get_numeric_data().copy()
         mn["little_float"] = np.array(12345.0, dtype="float16")
         mn["big_float"] = np.array(123456789101112.0, dtype="float64")
