+++ pandas/tests/groupby/test_groupby.py
@@ -136,18 +136,27 @@ def test_basic_aggregations(dtype):
         grouped.aggregate(lambda x: x * 2)
 
 
-def test_groupby_nonobject_dtype(mframe, df_mixed_floats):
-    key = mframe.index.codes[0]
-    grouped = mframe.groupby(key)
+def test_groupby_nonobject_dtype(multiindex_dataframe_random_data):
+    key = multiindex_dataframe_random_data.index.codes[0]
+    grouped = multiindex_dataframe_random_data.groupby(key)
     result = grouped.sum()
 
-    expected = mframe.groupby(key.astype("O")).sum()
+    expected = multiindex_dataframe_random_data.groupby(key.astype("O")).sum()
     assert result.index.dtype == np.int8
     assert expected.index.dtype == np.int64
     tm.assert_frame_equal(result, expected, check_index_type=False)
 
+
+def test_groupby_nonobject_dtype_mixed():
     # GH 3911, mixed frame non-conversion
-    df = df_mixed_floats.copy()
+    df = DataFrame(
+        {
+            "A": ["foo", "bar", "foo", "bar", "foo", "bar", "foo", "foo"],
+            "B": ["one", "one", "two", "three", "two", "two", "one", "three"],
+            "C": np.random.default_rng(2).standard_normal(8),
+            "D": np.array(np.random.default_rng(2).standard_normal(8), dtype="float32"),
+        }
+    )
     df["value"] = range(len(df))
 
     def max_value(group):
@@ -1052,7 +1061,7 @@ def test_raise_on_nuisance_python_multiple(three_group):
         grouped.mean()
 
 
-def test_empty_groups_corner(mframe):
+def test_empty_groups_corner(multiindex_dataframe_random_data):
     # handle empty groups
     df = DataFrame(
         {
@@ -1069,7 +1078,7 @@ def test_empty_groups_corner(mframe):
     expected = grouped.mean(numeric_only=True)
     tm.assert_frame_equal(result, expected)
 
-    grouped = mframe[3:5].groupby(level=0)
+    grouped = multiindex_dataframe_random_data[3:5].groupby(level=0)
     agged = grouped.apply(lambda x: x.mean())
     agged_A = grouped["A"].apply("mean")
     tm.assert_series_equal(agged["A"], agged_A)
@@ -1083,8 +1092,8 @@ def test_nonsense_func():
         df.groupby(lambda x: x + "foo")
 
 
-def test_wrap_aggregated_output_multindex(mframe):
-    df = mframe.T
+def test_wrap_aggregated_output_multindex(multiindex_dataframe_random_data):
+    df = multiindex_dataframe_random_data.T
     df["baz", "two"] = "peekaboo"
 
     keys = [np.array([0, 0, 1]), np.array([0, 0, 1])]
@@ -1103,24 +1112,24 @@ def test_wrap_aggregated_output_multindex(mframe):
         df.groupby(keys).aggregate(aggfun)
 
 
-def test_groupby_level_apply(mframe):
-    result = mframe.groupby(level=0).count()
+def test_groupby_level_apply(multiindex_dataframe_random_data):
+    result = multiindex_dataframe_random_data.groupby(level=0).count()
     assert result.index.name == "first"
-    result = mframe.groupby(level=1).count()
+    result = multiindex_dataframe_random_data.groupby(level=1).count()
     assert result.index.name == "second"
 
-    result = mframe["A"].groupby(level=0).count()
+    result = multiindex_dataframe_random_data["A"].groupby(level=0).count()
     assert result.index.name == "first"
 
 
-def test_groupby_level_mapper(mframe):
-    deleveled = mframe.reset_index()
+def test_groupby_level_mapper(multiindex_dataframe_random_data):
+    deleveled = multiindex_dataframe_random_data.reset_index()
 
     mapper0 = {"foo": 0, "bar": 0, "baz": 1, "qux": 1}
     mapper1 = {"one": 0, "two": 0, "three": 1}
 
-    result0 = mframe.groupby(mapper0, level=0).sum()
-    result1 = mframe.groupby(mapper1, level=1).sum()
+    result0 = multiindex_dataframe_random_data.groupby(mapper0, level=0).sum()
+    result1 = multiindex_dataframe_random_data.groupby(mapper1, level=1).sum()
 
     mapped_level0 = np.array(
         [mapper0.get(x) for x in deleveled["first"]], dtype=np.int64
@@ -1128,8 +1137,8 @@ def test_groupby_level_mapper(mframe):
     mapped_level1 = np.array(
         [mapper1.get(x) for x in deleveled["second"]], dtype=np.int64
     )
-    expected0 = mframe.groupby(mapped_level0).sum()
-    expected1 = mframe.groupby(mapped_level1).sum()
+    expected0 = multiindex_dataframe_random_data.groupby(mapped_level0).sum()
+    expected1 = multiindex_dataframe_random_data.groupby(mapped_level1).sum()
     expected0.index.name, expected1.index.name = "first", "second"
 
     tm.assert_frame_equal(result0, expected0)
