+++ pandas/tests/indexes/conftest.py
@@ -5,7 +5,6 @@ from pandas import (
     Series,
     array,
 )
-import pandas._testing as tm
 
 
 @pytest.fixture(params=[None, False])
@@ -38,24 +37,5 @@ def freq_sample(request):
 def listlike_box(request):
     """
     Types that may be passed as the indexer to searchsorted.
-    """
-    return request.param
-
-
-@pytest.fixture(
-    params=tm.ALL_REAL_NUMPY_DTYPES
-    + [
-        "object",
-        "category",
-        "datetime64[ns]",
-        "timedelta64[ns]",
-    ]
-)
-def any_dtype_for_small_pos_integer_indexes(request):
-    """
-    Dtypes that can be given to an Index with small positive integers.
-
-    This means that for any dtype `x` in the params list, `Index([1, 2, 3], dtype=x)` is
-    valid and gives the correct Index (sub-)class.
     """
     return request.param
