+++ pandas/tests/indexes/multi/test_duplicates.py
@@ -11,10 +11,29 @@ from pandas._libs import (
 from pandas import (
     NA,
     DatetimeIndex,
+    Index,
     MultiIndex,
     Series,
 )
 import pandas._testing as tm
+
+
+@pytest.fixture
+def idx_dup():
+    # compare tests/indexes/multi/conftest.py
+    major_axis = Index(["foo", "bar", "baz", "qux"])
+    minor_axis = Index(["one", "two"])
+
+    major_codes = np.array([0, 0, 1, 0, 1, 1])
+    minor_codes = np.array([0, 1, 0, 1, 0, 1])
+    index_names = ["first", "second"]
+    mi = MultiIndex(
+        levels=[major_axis, minor_axis],
+        codes=[major_codes, minor_codes],
+        names=index_names,
+        verify_integrity=False,
+    )
+    return mi
 
 
 @pytest.mark.parametrize("names", [None, ["first", "second"]])
