+++ pandas/tests/indexes/multi/test_formats.py
@@ -139,8 +139,11 @@ MultiIndex([...
            names=['first', ...], length=6)"""
             assert result == expected
 
-    def test_rjust(self, narrow_multi_index):
-        mi = narrow_multi_index
+    def test_rjust(self):
+        n = 1000
+        ci = pd.CategoricalIndex(list("a" * n) + (["abc"] * n))
+        dti = pd.date_range("2000-01-01", freq="s", periods=n * 2)
+        mi = MultiIndex.from_arrays([ci, ci.codes + 9, dti], names=["a", "b", "dti"])
         result = mi[:1].__repr__()
         expected = """\
 MultiIndex([('a', 9, '2000-01-01 00:00:00')],
@@ -182,8 +185,13 @@ MultiIndex([(  'a',  9, '2000-01-01 00:00:00'),
            names=['a', 'b', 'dti'], length=2000)"""
         assert result == expected
 
-    def test_tuple_width(self, wide_multi_index):
-        mi = wide_multi_index
+    def test_tuple_width(self):
+        n = 1000
+        ci = pd.CategoricalIndex(list("a" * n) + (["abc"] * n))
+        dti = pd.date_range("2000-01-01", freq="s", periods=n * 2)
+        levels = [ci, ci.codes + 9, dti, dti, dti]
+        names = ["a", "b", "dti_1", "dti_2", "dti_3"]
+        mi = MultiIndex.from_arrays(levels, names=names)
         result = mi[:1].__repr__()
         expected = """MultiIndex([('a', 9, '2000-01-01 00:00:00', '2000-01-01 00:00:00', ...)],
            names=['a', 'b', 'dti_1', 'dti_2', 'dti_3'])"""  # noqa: E501
