+++ pandas/tests/io/excel/test_readers.py
@@ -22,6 +22,7 @@ from pandas import (
     Index,
     MultiIndex,
     Series,
+    read_csv,
 )
 import pandas._testing as tm
 from pandas.core.arrays import (
@@ -115,6 +116,16 @@ def engine(engine_and_read_ext):
 def read_ext(engine_and_read_ext):
     engine, read_ext = engine_and_read_ext
     return read_ext
+
+
+@pytest.fixture
+def df_ref(datapath):
+    """
+    Obtain the reference data from read_csv with the Python engine.
+    """
+    filepath = datapath("io", "data", "csv", "test1.csv")
+    df_ref = read_csv(filepath, index_col=0, parse_dates=True, engine="python")
+    return df_ref
 
 
 def adjust_expected(expected: DataFrame, read_ext: str) -> None:
