+++ pandas/tests/io/json/conftest.py
@@ -7,10 +7,3 @@ def orient(request):
     Fixture for orients excluding the table format.
     """
     return request.param
-
-
-@pytest.fixture(params=["ujson", "pyarrow"])
-def engine(request):
-    if request.param == "pyarrow":
-        pytest.importorskip("pyarrow.json")
-    return request.param
