+++ pandas/tests/resample/conftest.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-import warnings
 
 import numpy as np
 import pytest
@@ -8,8 +7,6 @@ from pandas import (
     DataFrame,
     Series,
 )
-from pandas.core.indexes.datetimes import date_range
-from pandas.core.indexes.period import period_range
 
 # The various methods we support
 downsample_methods = [
@@ -42,40 +39,6 @@ def downsample_method(request):
 def resample_method(request):
     """Fixture for parametrization of Grouper resample methods."""
     return request.param
-
-
-@pytest.fixture
-def simple_date_range_series():
-    """
-    Series with date range index and random data for test purposes.
-    """
-
-    def _simple_date_range_series(start, end, freq="D"):
-        rng = date_range(start, end, freq=freq)
-        return Series(np.random.default_rng(2).standard_normal(len(rng)), index=rng)
-
-    return _simple_date_range_series
-
-
-@pytest.fixture
-def simple_period_range_series():
-    """
-    Series with period range index and random data for test purposes.
-    """
-
-    def _simple_period_range_series(start, end, freq="D"):
-        with warnings.catch_warnings():
-            # suppress Period[B] deprecation warning
-            msg = "|".join(["Period with BDay freq", r"PeriodDtype\[B\] is deprecated"])
-            warnings.filterwarnings(
-                "ignore",
-                msg,
-                category=FutureWarning,
-            )
-            rng = period_range(start, end, freq=freq)
-        return Series(np.random.default_rng(2).standard_normal(len(rng)), index=rng)
-
-    return _simple_period_range_series
 
 
 @pytest.fixture
