+++ pandas/tests/resample/test_datetime_index.py
@@ -53,6 +53,19 @@ def unit(request):
     return request.param
 
 
+@pytest.fixture
+def simple_date_range_series():
+    """
+    Series with date range index and random data for test purposes.
+    """
+
+    def _simple_date_range_series(start, end, freq="D"):
+        rng = date_range(start, end, freq=freq)
+        return Series(np.random.default_rng(2).standard_normal(len(rng)), index=rng)
+
+    return _simple_date_range_series
+
+
 def test_custom_grouper(index, unit):
     dti = index.as_unit(unit)
     s = Series(np.array([1] * len(dti)), index=dti, dtype="int64")
@@ -1206,14 +1219,6 @@ def test_corner_cases(unit):
     result = ts.resample("5min", closed="right", label="left").mean()
     ex_index = date_range("1999-12-31 23:55", periods=4, freq="5min").as_unit(unit)
     tm.assert_index_equal(result.index, ex_index)
-
-
-def test_corner_cases_period(simple_period_range_series):
-    # miscellaneous test coverage
-    len0pts = simple_period_range_series("2007-01", "2010-05", freq="M")[:0]
-    # it works
-    result = len0pts.resample("Y-DEC").mean()
-    assert len(result) == 0
 
 
 def test_corner_cases_date(simple_date_range_series, unit):
