+++ pandas/tests/strings/conftest.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 from pandas import Series
@@ -131,53 +130,3 @@ def any_string_method(request):
     ...     method(*args, **kwargs)
     """
     return request.param
-
-
-# subset of the full set from pandas/conftest.py
-_any_allowed_skipna_inferred_dtype = [
-    ("string", ["a", np.nan, "c"]),
-    ("bytes", [b"a", np.nan, b"c"]),
-    ("empty", [np.nan, np.nan, np.nan]),
-    ("empty", []),
-    ("mixed-integer", ["a", np.nan, 2]),
-]
-ids, _ = zip(*_any_allowed_skipna_inferred_dtype)  # use inferred type as id
-
-
-@pytest.fixture(params=_any_allowed_skipna_inferred_dtype, ids=ids)
-def any_allowed_skipna_inferred_dtype(request):
-    """
-    Fixture for all (inferred) dtypes allowed in StringMethods.__init__
-
-    The covered (inferred) types are:
-    * 'string'
-    * 'empty'
-    * 'bytes'
-    * 'mixed'
-    * 'mixed-integer'
-
-    Returns
-    -------
-    inferred_dtype : str
-        The string for the inferred dtype from _libs.lib.infer_dtype
-    values : np.ndarray
-        An array of object dtype that will be inferred to have
-        `inferred_dtype`
-
-    Examples
-    --------
-    >>> from pandas._libs import lib
-    >>>
-    >>> def test_something(any_allowed_skipna_inferred_dtype):
-    ...     inferred_dtype, values = any_allowed_skipna_inferred_dtype
-    ...     # will pass
-    ...     assert lib.infer_dtype(values, skipna=True) == inferred_dtype
-    ...
-    ...     # constructor for .str-accessor will also pass
-    ...     Series(values).str
-    """
-    inferred_dtype, values = request.param
-    values = np.array(values, dtype=object)  # object dtype to avoid casting
-
-    # correctness of inference tested in tests/dtypes/test_inference.py
-    return inferred_dtype, values
